unify list/grid fragments and adapters to use new framework
