+++ clients/android/NewsBlur/res/layout/view_story_row.xml
@@ -16,15 +16,15 @@
     android:layout_height="match_parent" >
 
     <View
-        android:id="@+id/row_item_favicon_borderbar_1"
+        android:id="@+id/story_item_favicon_borderbar_1"
         android:layout_width="5dp"
         android:layout_height="match_parent" />
 
     <View
-        android:id="@+id/row_item_favicon_borderbar_2"
+        android:id="@+id/story_item_favicon_borderbar_2"
         android:layout_width="5dp"
         android:layout_height="match_parent"
-        android:layout_toRightOf="@id/row_item_favicon_borderbar_1" />
+        android:layout_toRightOf="@id/story_item_favicon_borderbar_1" />
 
     <!--
         The next item to the right is actually the intel dot, but it has vertical bounds set relative to
@@ -35,7 +35,7 @@
     -->
 
     <ImageView
-        android:id="@+id/row_item_thumbnail"
+        android:id="@+id/story_item_thumbnail"
         android:layout_width="90dp"
         android:layout_height="90dp"
         android:layout_marginTop="1dp"
@@ -44,7 +44,7 @@
         android:visibility="gone" />
 
     <ImageView
-        android:id="@+id/row_item_feedicon"
+        android:id="@+id/story_item_feedicon"
         android:layout_width="18dp"
         android:layout_height="18dp"
         android:layout_marginTop="4dp"
@@ -52,24 +52,24 @@
         android:layout_alignParentLeft="true" />
 
     <TextView
-        android:id="@+id/row_item_feedtitle"
+        android:id="@+id/story_item_feedtitle"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginLeft="6dp"
         android:layout_marginTop="4dp"
-        android:layout_toRightOf="@id/row_item_feedicon"
-        android:layout_toLeftOf="@id/row_item_thumbnail"
+        android:layout_toRightOf="@id/story_item_feedicon"
+        android:layout_toLeftOf="@id/story_item_thumbnail"
         android:ellipsize="end"
         android:singleLine="true"
         style="?storyFeedTitleText" />
 
     <ImageView
-        android:id="@+id/row_item_saved_icon"
+        android:id="@+id/story_item_saved_icon"
         android:src="@drawable/clock"
         android:layout_width="20dp"
         android:layout_height="20dp"
-        android:layout_below="@id/row_item_feedtitle"
-        android:layout_toLeftOf="@id/row_item_thumbnail"
+        android:layout_below="@id/story_item_feedtitle"
+        android:layout_toLeftOf="@id/story_item_thumbnail"
         android:layout_alignWithParentIfMissing="true"
         android:layout_marginTop="2dp"
         android:layout_marginLeft="2dp"
@@ -77,12 +77,12 @@
         android:visibility="gone" />
 
     <ImageView
-        android:id="@+id/row_item_shared_icon"
+        android:id="@+id/story_item_shared_icon"
         android:src="@drawable/ic_shared"
         android:layout_width="20dp"
         android:layout_height="20dp"
-        android:layout_below="@id/row_item_feedtitle"
-        android:layout_toLeftOf="@id/row_item_saved_icon"
+        android:layout_below="@id/story_item_feedtitle"
+        android:layout_toLeftOf="@id/story_item_saved_icon"
         android:layout_alignWithParentIfMissing="true"
         android:layout_marginTop="2dp"
         android:layout_marginLeft="2dp"
@@ -90,11 +90,11 @@
         android:visibility="gone" />
 
     <TextView
-        android:id="@+id/row_item_title"
+        android:id="@+id/story_item_title"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_toLeftOf="@id/row_item_shared_icon"
-        android:layout_below="@id/row_item_feedicon"
+        android:layout_toLeftOf="@id/story_item_shared_icon"
+        android:layout_below="@id/story_item_feedicon"
         android:layout_marginLeft="33dp"
         android:paddingTop="6dp"
         android:paddingBottom="4dp"
@@ -103,11 +103,11 @@
         android:ellipsize="end" />
 
     <TextView
-        android:id="@+id/row_item_content"
+        android:id="@+id/story_item_content"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_toLeftOf="@id/row_item_thumbnail"
-        android:layout_below="@id/row_item_title"
+        android:layout_toLeftOf="@id/story_item_thumbnail"
+        android:layout_below="@id/story_item_title"
         android:layout_marginLeft="33dp"
         android:paddingBottom="4dp"
         android:paddingRight="4dp"
@@ -116,39 +116,39 @@
         style="?storySnippetText" />
 
     <TextView
-        android:id="@+id/row_item_date"
+        android:id="@+id/story_item_date"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_toLeftOf="@id/row_item_thumbnail"
+        android:layout_toLeftOf="@id/story_item_thumbnail"
         android:layout_alignWithParentIfMissing="true"
-        android:layout_below="@id/row_item_content"
+        android:layout_below="@id/story_item_content"
         android:paddingBottom="2dp"
         android:paddingRight="8dp"
         style="?storyDateText" />
 
     <TextView
-        android:id="@+id/row_item_author"
+        android:id="@+id/story_item_author"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_below="@id/row_item_content"
+        android:layout_below="@id/story_item_content"
         android:layout_alignParentLeft="true"
         android:layout_marginLeft="33dp"
-        android:layout_toLeftOf="@id/row_item_date"
+        android:layout_toLeftOf="@id/story_item_date"
         android:paddingBottom="2dp"
         android:ellipsize="end"
         android:singleLine="true"
         android:textColor="@color/story_author_text" />
 
     <RelativeLayout
-        android:id="@+id/row_item_inteldot_container"
+        android:id="@+id/story_item_inteldot_container"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:layout_alignTop="@id/row_item_title"
-        android:layout_alignBottom="@id/row_item_title"
-        android:layout_toRightOf="@id/row_item_favicon_borderbar_2" >
+        android:layout_alignTop="@id/story_item_title"
+        android:layout_alignBottom="@id/story_item_title"
+        android:layout_toRightOf="@id/story_item_favicon_borderbar_2" >
         
         <ImageView
-            android:id="@+id/row_item_inteldot"
+            android:id="@+id/story_item_inteldot"
             android:layout_width="9dp"
             android:layout_height="9dp"
             android:layout_marginLeft="6dp"
